+++ src/admin/KullaniciIslemleri/KullaniciSil.java
@@ -1,12 +1,10 @@
 package admin.KullaniciIslemleri;
 import java.sql.*;
 import java.util.ArrayList;
-import java.util.List;
 import java.util.Scanner;
-
 public class KullaniciSil {
 
-    public static List<Object> silinenlerListesi = new ArrayList<>(); //todo Listler classına taşımaya çalışalım
+
     public static void adminKullaniciSilMethodu() throws InterruptedException, SQLException, ClassNotFoundException {
         Class.forName("org.postgresql.Driver");
         Connection con = DriverManager.getConnection("jdbc:postgresql://localhost:5432/LibraryOtomation", "postgres", "1234");
@@ -15,7 +13,7 @@ public class KullaniciSil {
         System.out.println("Kullanici Silme Islemleri");
         String sqlKullanicilar = "SELECT * FROM kullanicilar";
         ResultSet kullanicilar =  st.executeQuery(sqlKullanicilar);
-        //1	"Ersin"	"AKUN"	"ersinakun34@gmail.com"	"1234"	"2126140853"	10
+        ArrayList<String> kullaniciIDList= new ArrayList<>();
         while (kullanicilar.next()) {
             System.out.println(kullanicilar.getInt(1)+"-"+ //id
                     kullanicilar.getString(2)+"-"+//ad
@@ -24,26 +22,31 @@ public class KullaniciSil {
                     kullanicilar.getString(5)+"-"+//şifre
                     kullanicilar.getString(6)+"-"+//tel
                     kullanicilar.getInt(7));//puan
+                    kullaniciIDList.add(String.valueOf((kullanicilar.getInt(1))));
         }
+
         Scanner scan = new Scanner(System.in);
         System.out.print("Silmek istediğiniz kullanıcının  ID numarısını giriniz :");
-        int secilenId = scan.nextInt();
+        String secilenId = scan.next();
 
+       if (kullaniciIDList.contains(secilenId)){
 
-        //todo olmayan bir id seçerse kontrol et, id dışında string girerse kontrol et. boşluk vs
-        String sqlKullaniciSilQuery = "DELETE FROM kullanicilar WHERE kullaniciid = ?";
-        PreparedStatement ps = con.prepareStatement(sqlKullaniciSilQuery);
+           String sqlKullaniciSilQuery = "DELETE FROM kullanicilar WHERE kullaniciid = ?";
+           PreparedStatement ps = con.prepareStatement(sqlKullaniciSilQuery);
+           ps.setInt(1, Integer.parseInt(secilenId));
+           ps.executeUpdate();
+           System.out.println(secilenId+ " id numaralı kullanıcı silindi");
 
-        ps.setInt(1,secilenId);
-        ps.executeUpdate();
-        System.out.println(secilenId+ " id numaralı kullanıcı silindi");
-
+       }
+       else {
+           System.out.println("Kayıt silinemedi/bulunamadı,lütfen tekrar deneyin!");
+            KullaniciMenusu.adminKullaniciIslemleriMenusuMethodu();
+       }
 
         con.close();
         st.close();
 
             KullaniciMenusu.adminKullaniciIslemleriMenusuMethodu();
-
         }
     }
